fix: disable CGO in tests to avoid NVML deprecation warnings
Problem:
- Test suite was failing with gotestfmt parser panic
- NVML library emits UTF-8 deprecation warnings during CGO compilation
- These warnings crashed the test output formatter

Solution:
1. Set CGO_ENABLED=0 in GitHub Actions test workflow
2. Split gpudetect into CGO/no-CGO builds:
   - detector_cgo.go: Full NVML support (production)
   - detector_nocgo.go: Stub returning CPU/Apple Silicon (tests)
3. Split tests:
   - detector_test.go: CGO-only (NVML unit tests)
   - detector_common_test.go: Works with/without CGO
4. Updated CLAUDE.md with CGO_ENABLED=0 test command

Impact:
- Tests run without NVML warnings
- Build still uses CGO for GPU detection in production
- No runtime behavior changes
- Faster test execution (no C compilation)
